backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.1
mysqlclient==2.2.6
djangorestframework-simplejwt==5.3.1
